Move yu-trinkle to seperate file
